Added non-functional password change garbage
Could not manage to get the devise edit passwords view to load for my
life. I tried creating a new controller/route and accessing it that way
but I couldn't get that to work either. Also I realize that the change
password link on the top is pretty trashy; we dont really have a
settings page so I didnt know where to put it for now. I will have time
to look at it more tomorrow.